I simplify hook by pushing logic to libs.
